requirements.txt
# Core Django
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# GraphQL
graphene-django>=3.1.0
graphql-core>=3.2.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
django-mptt>=0.14.0     # Tree structures

# Authentication & Security
django-allauth>=0.54.0
django-oauth-toolkit>=2.2.0
django-jwt-auth>=0.1.0

# File Handling & Media
Pillow>=9.5.0           # Image processing
django-storages>=1.13.0  # Cloud storage support

# History & Auditing
django-simple-history>=3.3.0
django-reversion>=6.0.0

# API & Serialization
django-filter>=23.0
drf-yasg>=1.21.0        # Swagger documentation

# Utilities
python-decouple>=3.8    # Environment variables
django-environ>=0.11.0
python-dateutil>=2.8.0

# Validation & Forms
django-phonenumber-field>=7.3.0
phonenumbers>=8.13.0

# Caching
redis>=4.5.0
django-redis>=5.2.0

# Background Tasks
celery>=5.3.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# Monitoring & Logging
django-debug-toolbar>=4.0.0
sentry-sdk>=1.25.0

# Testing
pytest>=7.3.0
pytest-django>=4.5.0
factory-boy>=3.3.0
coverage>=7.2.0

# Development Tools
black>=23.0.0           # Code formatting
flake8>=6.0.0           # Linting
isort>=5.12.0           # Import sorting
pre-commit>=3.3.0       # Git hooks

# Documentation
Sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# Healthcare Specific
fhirclient>=3.2.0       # FHIR integration
hl7>=0.4.4              # HL7 message parsing

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0         # Excel file handling

# Email & Notifications
django-anymail>=10.0
twilio>=8.0.0           # SMS notifications

# GIS & Location
django-geoposition>=0.3.0
geopy>=2.3.0

# PDF Generation
reportlab>=4.0.0
WeasyPrint>=59.0

# API Rate Limiting
django-ratelimit>=4.0.0

# Search
django-haystack>=3.2.0
elasticsearch>=8.8.0

# Health Checks
django-health-check>=3.16.0

# Production
gunicorn>=20.1.0        # WSGI server
whitenoise>=6.5.0       # Static files
django-extensions>=3.2.0

# Development Only (optional)
django-extensions>=3.2.0
django-debug-toolbar>=4.0.0 